# Windows/Arm64: Use 8.1 atomic instructions if they are available