--- a/src/coreclr/vm/codeman.cpp
+++ b/src/coreclr/vm/codeman.cpp
@@ -1575,6 +1575,10 @@ void EEJitManager::SetCpuInfo()
     }
 
 #endif // HOST_64BIT
+    if (CPUCompileFlags.IsSet(InstructionSet_Atomics))
+    {
+        g_atomic_present = true;
+    }
     if (GetDataCacheZeroIDReg() == 4)
     {
         // DCZID_EL0<4> (DZP) indicates whether use of DC ZVA instructions is permitted (0) or prohibited (1).