--- a/src/coreclr/vm/util.hpp
+++ b/src/coreclr/vm/util.hpp
@@ -25,6 +25,8 @@
 #define MAX_CACHE_LINE_SIZE 64
 #endif
 
+extern bool g_atomic_present;
+
 #ifndef TARGET_UNIX
 // Copied from malloc.h: don't want to bring in the whole header file.
 void * __cdecl _alloca(size_t);
@@ -71,6 +73,76 @@ BOOL inline FitsInU4(unsigned __int64 val)
     return val == (unsigned __int64)(unsigned __int32)val;
 }
 
+
+#if defined(TARGET_ARM64)
+
+FORCEINLINE LONG  FastInterlockedCompareExchange(
+    LONG volatile *Destination,
+    LONG Exchange,
+    LONG Comperand)
+{
+    printf("g_atomic_present (FastInterlockedCompareExchange)= %d\n", g_atomic_present);
+    if (g_atomic_present)
+    {
+        return (LONG) __casal32((unsigned __int32*) Destination, (unsigned  __int32)Comperand, (unsigned __int32)Exchange);
+    }
+    else
+    {
+        return InterlockedCompareExchange(Destination, Exchange, Comperand);
+    }
+}
+
+FORCEINLINE LONGLONG  FastInterlockedCompareExchange64(
+    IN OUT LONGLONG volatile *Destination,
+    IN LONGLONG Exchange,
+    IN LONGLONG Comperand)
+{
+    printf("g_atomic_present (FastInterlockedCompareExchange64)= %d\n", g_atomic_present);
+    if (g_atomic_present)
+    {
+        return (LONGLONG) __casal64((unsigned __int64*) Destination, (unsigned  __int64)Comperand, (unsigned __int64)Exchange);
+    }
+    else
+    {
+        return InterlockedCompareExchange64(Destination, Exchange, Comperand);
+    }
+}
+
+FORCEINLINE LONG FastInterlockedCompareExchangeAcquire(
+  IN OUT LONG volatile *Destination,
+  IN LONG Exchange,
+  IN LONG Comperand
+)
+{
+    printf("g_atomic_present (FastInterlockedCompareExchangeAcquire)= %d\n", g_atomic_present);
+    if (g_atomic_present)
+    {
+        return (LONG) __casa32((unsigned __int32*) Destination, (unsigned  __int32)Comperand, (unsigned __int32)Exchange);
+    }
+    else
+    {
+        return InterlockedCompareExchangeAcquire(Destination, Exchange, Comperand);
+    }
+}
+
+FORCEINLINE LONG FastInterlockedCompareExchangeRelease(
+  IN OUT LONG volatile *Destination,
+  IN LONG Exchange,
+  IN LONG Comperand
+)
+{
+    printf("g_atomic_present (FastInterlockedCompareExchangeRelease)= %d\n", g_atomic_present);
+    if (g_atomic_present)
+    {
+        return (LONG) __casl32((unsigned __int32*) Destination, (unsigned  __int32)Comperand, (unsigned __int32)Exchange);
+    }
+    else
+    {
+        return InterlockedCompareExchangeRelease(Destination, Exchange, Comperand);
+    }
+}
+
+#endif // TARGET_ARM64
 
 
 //************************************************************************