@jkotas 2022-06-17
> It is little puzzling why 16 threads would just show slight performance improvement and I will check with few people offline.

It may be caused by threads finishing the work before all threads are running. It is also why you do not see any improvement for 2 threads - there should be improvement for 2 threads too. You may want to change the spin count to `1_000_000` and add Barrier in frond the loop so that all threads wait for each other getting created before starting the work.