@kunalspathak 2022-06-18
Thanks @jkotas . I updated to `100000` for shorter test run duration and also added memory barrier before starting the threads. Here is modified code along with test result. I think they all make sense now except the Min/2 threads, but if you see raw numbers, the variance for `main` is around 64%, so I won't trust that number.


```c#
namespace Monitor_Lock  
{  
    class Program  
    {  
        static readonly object _object = new object();  
        static long result = 0;
  
        static void TestLock()  
        {
            for (int i = 0; i < 100000; i++)
            {
                lock (_object)  
                {  
                    result += i;
                }
            }
        }  
  
        static void Main(string[] args)      
        {
            Stopwatch sw = new Stopwatch();
            int count = int.Parse(args[0]);
            List<Thread> threads = new List<Thread>();
            sw.Start();
            for (int i = 0; i < count; i++)  
            {  
                ThreadStart start = new ThreadStart(TestLock);  
                Thread thread = new Thread(start);
                threads.Add(thread);
            }

            Interlocked.MemoryBarrier();

            for (int i = 0; i < count; i++)  
            {  
                threads[i].Start();
            }

            for (int i = 0; i < count; i++)
            {
                threads[i].Join();
            }
            sw.Stop();
            Console.WriteLine($"{count} threads took {sw.Elapsed.TotalMilliseconds} msec.");
        }
    }  
}  
```
Average:

| Threads | main     | PR       | diff    |
|---------|----------|----------|---------|
| 1       | 16.74636 | 14.06296 | -16.02% |
| 2       | 18.3666  | 16.4355  | -10.51% |
| 4       | 43.69844 | 43.00226 | -1.59%  |
| 8       | 86.08646 | 77.678   | -9.77%  |
| 16      | 200.1058 | 167.4846 | -16.30% |
| 32      | 342.4692 | 308.6021 | -9.89%  |
| 64      | 605.3218 | 542.0234 | -10.46% |
| 80      | 807.5776 | 687.6071 | -14.86% |
| 128     | 1230.828 | 1167.448 | -5.15%  |
| 256     | 2674.654 | 2251.305 | -15.83% |

Min:
| Threads | main     | PR       | diff    |
|---------|----------|----------|---------|
| 1       | 5.1449   | 5.0627   | -1.60%  |
| 2       | 11.7118  | 15.5086  | 32.42%  |
| 4       | 36.8293  | 27.3163  | -25.83% |
| 8       | 74.6232  | 56.595   | -24.16% |
| 16      | 140.0395 | 132.3511 | -5.49%  |
| 32      | 325.5867 | 249.9301 | -23.24% |
| 64      | 578.1493 | 453.8749 | -21.50% |
| 80      | 700.3981 | 562.5034 | -19.69% |
| 128     | 1143.339 | 1060.977 | -7.20%  |
| 256     | 2413.856 | 2139.666 | -11.36% |

Max:

| Threads | main     | PR       | diff    |
|---------|----------|----------|---------|
| 1       | 5.6196   | 5.6733   | 0.96%   |
| 2       | 31.1041  | 17.2587  | -44.51% |
| 4       | 61.231   | 61.2019  | -0.05%  |
| 8       | 102.0049 | 100.7827 | -1.20%  |
| 16      | 245.4816 | 204.5831 | -16.66% |
| 32      | 354.2334 | 359.8534 | 1.59%   |
| 64      | 628.9136 | 601.3752 | -4.38%  |
| 80      | 1041.367 | 799.7853 | -23.20% |
| 128     | 1336.44  | 1259.636 | -5.75%  |
| 256     | 2803.754 | 2360.407 | -15.81% |

<details>
<summary> Raw numbers of all iterations </summary>

| Threads | main     | PR       |
|---------|----------|----------|
| 1       | 5.6196   | 5.6733   |
| 1       | 5.1449   | 5.0627   |
| 1       | 5.1681   | 5.1654   |
| 1       | 5.1943   | 5.1351   |
| 1       | 5.3091   | 5.2476   |
| 2       | 13.2151  | 16.3577  |
| 2       | 11.7118  | 15.9422  |
| 2       | 22.3917  | 17.2587  |
| 2       | 31.1041  | 15.5086  |
| 2       | 13.4103  | 17.1103  |
| 4       | 43.1938  | 35.0976  |
| 4       | 37.7045  | 52.4248  |
| 4       | 39.5336  | 61.2019  |
| 4       | 61.231   | 27.3163  |
| 4       | 36.8293  | 38.9707  |
| 8       | 79.4986  | 56.595   |
| 8       | 74.6232  | 69.2368  |
| 8       | 78.4101  | 83.1975  |
| 8       | 102.0049 | 100.7827 |
| 8       | 95.8955  | 78.578   |
| 16      | 241.5675 | 193.9641 |
| 16      | 148.6181 | 132.3511 |
| 16      | 224.8223 | 170.8608 |
| 16      | 245.4816 | 135.664  |
| 16      | 140.0395 | 204.5831 |
| 32      | 341.4042 | 249.9301 |
| 32      | 325.5867 | 343.8633 |
| 32      | 338.0256 | 359.8534 |
| 32      | 354.2334 | 315.5561 |
| 32      | 353.0962 | 273.8075 |
| 64      | 628.9136 | 569.8978 |
| 64      | 578.1493 | 566.4561 |
| 64      | 587.394  | 453.8749 |
| 64      | 627.6597 | 601.3752 |
| 64      | 604.4926 | 518.5128 |
| 80      | 831.2326 | 571.1818 |
| 80      | 700.3981 | 755.7222 |
| 80      | 1041.367 | 562.5034 |
| 80      | 732.4616 | 799.7853 |
| 80      | 732.4285 | 748.8428 |
| 128     | 1336.44  | 1232.302 |
| 128     | 1219.41  | 1060.977 |
| 128     | 1174.404 | 1259.636 |
| 128     | 1280.545 | 1181.896 |
| 128     | 1143.339 | 1102.428 |
| 256     | 2775.944 | 2139.666 |
| 256     | 2413.856 | 2265.159 |
| 256     | 2703.091 | 2194.7   |
| 256     | 2803.754 | 2296.592 |
| 256     | 2676.625 | 2360.407 |
</details>