@jkotas 2022-06-18
>  also added memory barrier before starting the threads.

Nit: I meant `Barrier` not memory barrier. Something like:
```
using System.Threading;
using System.Diagnostics;

class Program
{
    static readonly object _object = new object();
    static long result = 0;
    static int _iterations;
    static Barrier _barrier;

    static void TestLock()
    {
        for (;;)
        {
            _barrier.SignalAndWait();
            int iterations = _iterations;
            for (int i = 0; i < iterations; i++)
            {
                lock (_object)
                {
                    result += i;
                }
            }
            _barrier.SignalAndWait();
        }
    }

    static void Main(string[] args)
    {
        Stopwatch sw = new Stopwatch();
        int count = int.Parse(args[0]);
        _iterations = 10_000_000 / count;
        _barrier = new Barrier(count + 1);
        for (int i = 0; i < count; i++)
        {
            new Thread(TestLock) { IsBackground = true }.Start();
        }

        for (int i = 0; i < 5; i++)
        {
            _barrier.SignalAndWait();
            sw.Restart();
            _barrier.SignalAndWait();
            if (i != 0) // Throw away warm up run
               Console.WriteLine($"{count} threads took {sw.Elapsed.TotalMilliseconds} msec.");
        }
    }
}
```